backend/requirements/dev.txt
-r base.txt
django-debug-toolbar==3.2  # Конфигурируемая панель с информацией для отладки проекта